+++ cdlclient/__init__.py
@@ -17,7 +17,7 @@ application through XML-RPC protocol.
 from cdlclient.baseproxy import SimpleBaseProxy  # noqa: F401
 from cdlclient.remote import SimpleRemoteProxy  # noqa: F401
 
-__version__ = "0.8.1"
+__version__ = "0.9.0"
 __required_server_version__ = "0.11.0"
 __docurl__ = "https://cdlclient.readthedocs.io/en/latest/"
 __homeurl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/"